Dodavanje početnih fajlova na branch izvještaja
